refactor: bound long polling metrics
(cherry picked from commit 7c569afb872dd220a0129ba574583389c4a8848a)
